correlator: moved to accumulators.
