+++ src/script_interface/CMakeLists.txt
@@ -4,7 +4,6 @@ set(EspressoScriptInterface_SRC
     "${CMAKE_CURRENT_SOURCE_DIR}/collision_detection/initialize.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/constraints/initialize.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/cluster_analysis/initialize.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/correlators/initialize.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/lbboundaries/initialize.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/virtual_sites/initialize.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/observables/initialize.cpp"
