+++ src/script_interface/accumulators/AutoUpdateCorrelators.hpp
@@ -24,24 +24,24 @@
 #include "Correlator.hpp"
 #include "ScriptInterface.hpp"
 #include "ScriptObjectRegistry.hpp"
-#include "core/correlators.hpp"
+#include "core/accumulators.hpp"
 
 namespace ScriptInterface {
-namespace Correlators {
+namespace Accumulators {
 
 class AutoUpdateCorrelators : public ScriptObjectRegistry<Correlator> {
   virtual void add_in_core(std::shared_ptr<Correlator> obj_ptr) override {
     obj_ptr->correlator()->start_auto_update();
-    ::Correlators::auto_update_correlators.push_back(obj_ptr->correlator());
+    ::Accumulators::auto_update_correlators.push_back(obj_ptr->correlator());
   }
   virtual void remove_in_core(std::shared_ptr<Correlator> obj_ptr) override {
-    auto it = std::find(::Correlators::auto_update_correlators.begin(),
-                        ::Correlators::auto_update_correlators.end(),
+    auto it = std::find(::Accumulators::auto_update_correlators.begin(),
+                        ::Accumulators::auto_update_correlators.end(),
                         obj_ptr->correlator());
 
-    if (it != ::Correlators::auto_update_correlators.end()) {
+    if (it != ::Accumulators::auto_update_correlators.end()) {
       obj_ptr->correlator()->stop_auto_update();
-      ::Correlators::auto_update_correlators.erase(it);
+      ::Accumulators::auto_update_correlators.erase(it);
 
     } else {
       throw "Could not find Correlator to remove";
