--- src/script_interface/correlators/initialize.hpp
@@ -1,31 +1,0 @@
-/*
-  Copyright (C) 2015,2016 The ESPResSo project
-
-  This file is part of ESPResSo.
-
-  ESPResSo is free software: you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation, either version 3 of the License, or
-  (at your option) any later version.
-
-  ESPResSo is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU General Public License for more details.
-
-  You should have received a copy of the GNU General Public License
-  along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
-#ifndef SCRIPT_INTERFACE_CORRELATORS_INITIALIZE_HPP
-#define SCRIPT_INTERFACE_CORRELATORS_INITIALIZE_HPP
-
-namespace ScriptInterface {
-namespace Correlators {
-
-void initialize();
-
-} /* namespace Correlators */
-} /* namespace ScriptInterface */
-
-#endif
